LayoutTests/imported/w3c/web-platform-tests/css/css-transitions/after-change-style-inherited-try-fallback-expected.txt


FAIL Position fallback correctly applied to after-change style assert_equals: #inner halfway between black and lime expected "rgb(0, 128, 0)" but got "rgb(64, 64, 0)"
